+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix URL not updating while navigating.
 * Fix display of faction levels.
 * Fix The Drifter showing an error because of a redacted item.
+* Fix a case where the Google Drive data file would not be created.
 
 # 5.3.1 (2018-11-26)